+++ Fantasy/game/system/Time.h
@@ -6,9 +6,9 @@ public:
     //程序运行时间(ms)
     static long time;
     //每帧时间增量
-    static double deltaTime;
+    static float deltaTime;
     //时间速度缩放
-    static double timeScale;
+    static float timeScale;
 
     static void tick();
 };
